+++ spinta/cli/helpers/store.py
@@ -3,6 +3,8 @@ from pathlib import Path
 import click
 
 from spinta import commands
+from spinta.auth import auth_server_keys_exists
+from spinta.auth import client_exists
 from spinta.auth import create_client_file
 from spinta.auth import gen_auth_server_keys
 from spinta.components import Context
@@ -10,23 +12,43 @@ from spinta.components import Store
 from spinta.core.config import DEFAULT_CONFIG_PATH
 
 
-def _ensure_config_dir(path: Path, *, verbose: bool = True):
+def _ensure_config_dir(
+    path: Path,
+    default_auth_client: str,
+    *,
+    verbose: bool = True,
+):
     """Create default config_path directory if it does not exits"""
-    if path.exists() or path != DEFAULT_CONFIG_PATH:
-        return
 
-    if verbose:
-        click.echo(f"Initializing default config dir: {path}")
+    # If a non default config directory is given, it should exist.
+    if not path.exists() and path != DEFAULT_CONFIG_PATH:
+        raise Exception(f"Config dir {path} does not exist!")
 
-    path.mkdir(parents=True)
-    (path / 'clients').mkdir(parents=True)
-    gen_auth_server_keys(path)
-    create_client_file(path / 'clients', 'default', secret=None, scopes=[
-        'spinta_getall',
-        'spinta_getone',
-        'spinta_search',
-        'spinta_changes',
-    ])
+    # Ensure clients directory.
+    clients_path = (path / 'clients')
+    clients_path.mkdir(parents=True, exist_ok=True)
+
+    # Ensure auth server keys.
+    if not auth_server_keys_exists(path):
+        if verbose:
+            click.echo(f"Initializing auth server keys: {path}")
+        gen_auth_server_keys(path)
+
+    # Ensure default client.
+    if not client_exists(clients_path, default_auth_client):
+        if verbose:
+            click.echo(f"Initializing default auth client: {path}")
+        create_client_file(
+            clients_path,
+            default_auth_client,
+            secret=None,
+            scopes=[
+                'spinta_getall',
+                'spinta_getone',
+                'spinta_search',
+                'spinta_changes',
+            ],
+        )
 
 
 def load_store(
@@ -38,7 +60,11 @@ def load_store(
     config = context.get('config')
     commands.load(context, config)
     if ensure_config_dir:
-        _ensure_config_dir(config.config_path, verbose=verbose)
+        _ensure_config_dir(
+            config.config_path,
+            config.default_auth_client,
+            verbose=verbose,
+        )
     commands.check(context, config)
     store = context.get('store')
     commands.load(context, store)
@@ -49,7 +75,7 @@ def prepare_manifest(
     context: Context,
     *,
     verbose: bool = True,
-    ensure_config_dir: bool = False
+    ensure_config_dir: bool = False,
 ) -> Store:
     store = load_store(
         context,